cm, cm_net, evse_common: Finish charge mode change request impl.

+++ software/src/modules/charge_manager/charge_manager.cpp
@@ -345,7 +345,7 @@ void ChargeManager::start_manager_task()
         auto &charger_alloc = this->charger_allocation_state[i];
         auto charge_mode = this->cm_to_config_cm(this->charger_state[i].charge_mode);
 
-        if(cm_networking.send_manager_update(i, charger_alloc.allocated_current, charger_alloc.cp_disconnect, charger_alloc.allocated_phases, static_cast<uint8_t>(charge_mode), this->supported_charge_mode_bitmask))
+        if(cm_networking.send_manager_update(i, charger_alloc.allocated_current, charger_alloc.cp_disconnect, charger_alloc.allocated_phases, charge_mode, this->supported_charge_mode_bitmask))
             ++i;
 
     }, cm_send_delay);
